+++ BuildVRFIndex.py
@@ -31,14 +31,16 @@ from argparse					import ArgumentParser, RawDescriptionHelpFormatter
 from base64						import b64decode
 from ConfigParser				import ConfigParser
 from datetime                   import datetime
-from Exscript                   import Account, Queue, Host
+from Exscript                   import Account, Queue, Host, Logger
 from Exscript.protocols 		import SSH2
 from Exscript.util.file			import get_hosts_from_file
+from Exscript.util.log          import log_to
 from Exscript.util.decorator    import autologin
 from Exscript.util.interact     import read_login
+from Exscript.util.report		import status,summarize
 from re							import sub
 from sys						import stdout
-from os							import name, remove, system
+from os							import makedirs, name, path, remove, system
 
 
 class Application:
@@ -46,13 +48,15 @@ class Application:
 # details across all my applications.  Also used to display information when
 # application is executed with "--help" argument.
 	author = "Aaron Melton <aaron@aaronmelton.com>"
-	date = "(2013-08-28)"
+	date = "(2013-08-29)"
 	description = "Creates a CSV of VRF Names across multiple Cisco routers"
 	name = "BuildVRFIndex.py"
 	url = "https://github.com/aaronmelton/BuildVRFIndex"
-	version = "v0.0.6-alpha"
+	version = "v0.0.7-alpha"
 
 
+logger = Logger()	# Log stuff
+@log_to(logger)		# Logging decorator; Must precede buildIndex!
 @autologin()		# Exscript login decorator; Must precede buildIndex!
 def buildIndex(job, host, socket):
 # This function builds the index file by connecting to the router and extracting all
@@ -156,13 +160,34 @@ def routerLogin():
 		else:							# Else use username/password from configFile
 			account = Account(name=username, password=b64decode(password))
 		
-		queue = Queue(verbose=0, max_threads=1)	# Minimal message from queue, 1 threads
+		# Minimal message from queue, 1 threads
+		queue = Queue(verbose=0, max_threads=1, stderr=(open(os.devnull, 'w')))
 		queue.add_account(account)				# Use supplied user credentials
 		print
 		stdout.write("--> Building index...") 	# Print without trailing newline
 		queue.run(hosts, buildIndex)			# Create queue using provided hosts
 		queue.shutdown()						# End all running threads and close queue
 		
+		# Define log filename based on date
+		logFilename = logFileDirectory+'BuildVRFIndex_'+date+'.log'
+
+		# Check to see if logFilename currently exists.  If it does, append an
+		# integer onto the end of the filename until logFilename no longer exists
+		incrementLogFilename = 1
+		while fileExist(logFilename):
+			logFilename = logFileDirectory+'BuildVRFIndex_'+date+'_'+str(incrementLogFilename)+'.log'
+			incrementLogFilename = incrementLogFilename + 1
+
+		# Write log results to logFile
+		with open(logFilename, 'w') as outputLogFile:
+			try:
+				outputLogFile.write(summarize(logger))
+
+			# Exception: router file was not able to be opened
+			except IOError:
+				print
+				print "--> An error occurred opening "+logFileDirectory+logFile+"."
+
 	# Exception: router file could not be opened
 	except IOError:
 		print
@@ -205,7 +230,7 @@ except IOError:
 			print "--> Config file not found; Creating "+configFile+"."
 			exampleFile.write("## BuildVRFIndex.py CONFIGURATION FILE ##\n#\n")
 			exampleFile.write("[account]\n#password is base64 encoded! Plain text passwords WILL NOT WORK!\n#Use website such as http://www.base64encode.org/ to encode your password\nusername=\npassword=\n#\n")
-			exampleFile.write("[BuildVRFIndex]\n#Check your paths! Files will be created; Directories will not.\n#Bad directories may result in errors!\n#variable=C:\path\\to\\filename.ext\nrouterFile=routers.txt\nindexFile=index.txt\nindexFileTmp=index.txt.tmp\n")
+			exampleFile.write("[BuildVRFIndex]#variable=C:\path\\to\\filename.ext\nrouterFile=routers.txt\nindexFile=index.txt\nindexFileTmp=index.txt.tmp\nlogFileDirectory=\n")
 
 	# Exception: file could not be created
 	except IOError:
@@ -222,7 +247,18 @@ finally:
 	routerFile = config.get('BuildVRFIndex', 'routerFile')
 	indexFile = config.get('BuildVRFIndex', 'indexFile')
 	indexFileTmp = config.get('BuildVRFIndex', 'indexFileTmp')
+	logFileDirectory = config.get('BuildVRFIndex', 'logFileDirectory')
+
+	# If logFileDirectory does not contain trailing backslash, append one
+	if logFileDirectory != '':
+		if logFileDirectory[-1:] != "\\":
+			logFileDirectory = logFileDirectory+"\\"
+			if not path.exists(logFileDirectory): makedirs(logFileDirectory)
 	
+	# Define 'date' variable for use in the output filename
+	date = datetime.now()	# Determine today's date
+	date = date.strftime('%Y%m%d')	# Format date as YYYYMMDD
+
 	if fileExist(routerFile):
 	# If the routerFile exists, proceed to login to routers
 		if fileExist(indexFile):	# If indexFile exists